requirements.txt
fado==2.0.4
pyyaml>=6.0